auto-fill: try to hide connecting stitches in the fill pattern
This lets us auto-fill regions with holes in it.  Just cut the region exactly
along the fill angle to convert a region with a hole to a region without.

This:
 __________________
|                  |
|      ______      |
|     |      |     |
|     |      |     |
|     |      |     |
|     |______|     |
|__________________|

Becomes this:
 __________________
|                  |
|      ______      |
|     |      |     |
|     |      |     |
 =====       |     |
|     |______|     |
|__________________|

The cut ( ===== ) is less than the width of the row spacing.  Any connecting
stitch that follows this path becomes hidden among the rest of the fill --
an extra row of stitching isn't visible so long as the staggers line up.