switch stitch view from paths to polylines
Polylines in SVG are essentially the same thing as paths, but with no curves.
But inkscape-embroidery treats polylines as literal stitches; it does not
do satins, fills, etc but instead takes the coordinates and uses them directly
as stitches.

The advantage in this case is that one can re-run the extension on its own
output and get back the same stitch output.  Why would you want to do this?

Consider the case where you have a basic embroidery machine that can only do
a 4x4" square.  If you have something that's a bit bigger than 4 inches wide,
it might still fit in the square if you turn it at an angle.  But working with
your entire design rotated at an angle is a total pain, and you'd also have to
rotate the angles of all the fills to compensate.

Instead, just run the extension, rotate the stitches to fit, and re-run, and
your stitch file will work on your 4x4" machine.