parse and use viewBox attribute
The viewBox effectively adds global scaling and translation to all shapes in
the SVG.  Borrowing from inkscape-silhouette, we construct a transform from
the viewBox and apply it to all objects.

When adding the stitch plan into the SVG, we need to compensate for this
implied transformation, which we do by adding its inverse as a transform
on the stitch plan polylines.

All of this allows us to do away with the nonstandard 10 pixels per mm that
was previously hardcoded into inkstitch.  Old designs can add a viewBox to
switch from 10 pixels per mm to the standard 96 ppi that Inkscape uses.