fix RUNPATH for wx libraries
For reasons unclear to me, the wx shared libraries have their RUNPATH set
to the library path of my _system_ python (not even the virtualenv python).
That's not going to work, especially when we package up and ship the
virtualenv.

Setting the RUNPATH to $ORIGIN says that the linker should look for
libwx_gtk3u_core-3.0.so.0 et al. in the same directory as _core.so.