get rid of antiquated notion of stitch params passed as command line args
Way back in the day, the only way to control things like the stitch length
or satin density was through global options specified in the extension
settings.  We've long sinced moved to per-object params, and for backward
compatibility, we defaulted to the command-line arguments.

That means that it was possible to get different stitch reslts from the
same SVG file if you passed different settings.  For that reason, I
never touched mine.  I didn't intend for my users to at all.

At this point, the command line arguments are just getting in the way of
implementing more features, so I'm getting rid of them and moving the
defaults into the parameters system.