v1.8.0: simplified extension packaging
*Please fully delete Ink/Stitch from your extensions directory before
installing this version.*

Ink/Stitch was previously a set of separate scripts, each implementing an
extension in a one-to-one correspondence.  Now all of the extensions are
implemented by a single python script.  This is an "under-the-hood" only
change that should not have any noticeable effects.

There is one big visible improvement: the zip for Linux is one third of
the previous size!