new extension: split satin (#315)
This branch adds a new command to split a satin column at a specified point.  The split happens at a stitch boundary to ensure that the two resulting satins sew just like the original.  All parameters set on the original satin remain set on the two new satins, and all rungs are retained.  If one of the satins would have no rungs left, a new rung is added.

How to use:

1. Select a satin column (simple satin doesn't work)
2. Attach the "Satin split point" command using the "Attach commands to selected objects" extension.
3. Move the symbol (or just the connector line's endpoint) to point to the exact spot you want the satin to be split at.
4. Select the satin column again.
5. Run "Split Satin Column".
6. The split point command and connector line disappear, and nothing else appears to have happened.  Select your satin and you'll see that it's been split.

This extension is a by-product of my initial work on #214.  Ink/Stitch will need the ability to split a satin at an arbitrary point, and I figured, why not go ahead and release that functionality as an extension while I'm at it? :)