v1.25.0: Electron for Print PDF UI
This version transitions Print PDF to use Electron, a bundled version of
Chromium browser.  The Print PDF interface no longer opens in your default
web browser.  This change shouldn't be particularly significant for most users.

This lays the groundwork for converting the rest of Ink/Stitch's GUI to use
Electron, which will enable faster development and a more modern interface.

The Ink/Stitch installation ZIP/tarball is significantly bigger in this
version.  Once we've finished the work of converting the existing interface
to Electron, the ZIP/tarball will shrink.