Update for Inkstitch Sulky Rayon Thread palette
I don't remember from where but I have a file "RGB-Liste SULKY RAYON 40.xlsx" with RGB values for all threads currently available from Sulky as Rayon 40 threads.
The order in the file is a bit different from the current one. This is the file converted to a gpl palette. It has 9 additional colors.