fix manually clicking pencil icon
The pencil icon was originally added in #301, and my explanation for what it does is there.

Clicking the pencil icon _should_ tell Ink/Stitch that you want that setting to be written to all relevant selected objects, even though you haven't actually changed the value.  It should also cause that setting to be included if you save a preset.  That's been broken for awhile now (see #3888).

This PR should fix that and make clicking the pencil icon work again.